reporting list of campaigns
