--- app/controllers/reports.js
@@ -1,3 +1,0 @@
-exports.generateReport = () => {
-  console.log('reports.js generate report');
-}
